backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pymongo==4.7.3
python-dotenv==1.0.1
bcrypt==4.1.3
PyJWT==2.8.0
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2
requests==2.32.3
python-dateutil==2.9.0
